fasgfdgsag

+++ MeteorOS/SystemX86/CoreApps/shell.lua
@@ -76,9 +76,14 @@ local function installUpdate()
     shell.run("wget https://raw.githubusercontent.com/Mag1cpunch/MeteorOS-AR2/main/MeteorOS/Programs/orescanner.lua /MeteorOS/Programs/orescanner.lua")
     shell.run("wget https://raw.githubusercontent.com/Mag1cpunch/MeteorOS-AR2/main/MeteorOS/Programs/arterm.lua /MeteorOS/Programs/arterm.lua")
     shell.run("wget https://energetic.pw/computercraft/ore3d/assets/ore3d.lua /MeteorOS/Programs/ore3d.lua")
-    fs.delete("/MeteorOS/SystemX86/APIS")
-    shell.run("mkdir /MeteorOS/SystemX86/APIS")
-    shell.run("wget https://raw.githubusercontent.com/Mag1cpunch/MeteorOS-AR2/main/MeteorOS/SystemX86/APIS/ar_terminal.lua /MeteorOS/SystemX86/APIS/ar_terminal.lua")
+    if fs.exists("/MeteorOS/SystemX86/APIS") then
+        fs.delete("/MeteorOS/SystemX86/APIS")
+        shell.run("mkdir /MeteorOS/SystemX86/APIS")
+        shell.run("wget https://raw.githubusercontent.com/Mag1cpunch/MeteorOS-AR2/main/MeteorOS/SystemX86/APIS/ar_terminal.lua /MeteorOS/SystemX86/APIS/ar_terminal.lua")
+    else
+        shell.run("mkdir /MeteorOS/SystemX86/APIS")
+        shell.run("wget https://raw.githubusercontent.com/Mag1cpunch/MeteorOS-AR2/main/MeteorOS/SystemX86/APIS/ar_terminal.lua /MeteorOS/SystemX86/APIS/ar_terminal.lua")
+    end
     os.reboot()
 end
 local function listGitFiles(username, repo, path)
